Move Google extractor contract test
